@@ -46,7 +46,7 @@ dispatch_threshold_local_mean = Dispatcher(name="dispatch_threshold_local_mean")
 # ================== convolve ==================
 @dispatch_convolve.register(np.ndarray)
 def numpy_convolve(*args, **kwargs):
-    return scipy.ndimage.filters.convolve
+    return scipy.ndimage.convolve
 
 
 @dispatch_convolve.register_lazy("cupy")
@@ -56,13 +56,13 @@ def register_cupy_convolve():
 
     @dispatch_convolve.register(cupy.ndarray)
     def cupy_convolve(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.convolve
+        return cupyx.scipy.ndimage.convolve
 
 
 # ================== correlate ==================
 @dispatch_correlate.register(np.ndarray)
 def numpy_correlate(*args, **kwargs):
-    return scipy.ndimage.filters.correlate
+    return scipy.ndimage.correlate
 
 
 @dispatch_correlate.register_lazy("cupy")
@@ -72,13 +72,13 @@ def register_cupy_correlate():
 
     @dispatch_correlate.register(cupy.ndarray)
     def cupy_correlate(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.correlate
+        return cupyx.scipy.ndimage.correlate
 
 
 # ================== laplace ==================
 @dispatch_laplace.register(np.ndarray)
 def numpy_laplace(*args, **kwargs):
-    return scipy.ndimage.filters.laplace
+    return scipy.ndimage.laplace
 
 
 @dispatch_laplace.register_lazy("cupy")
@@ -88,13 +88,13 @@ def register_cupy_laplace():
 
     @dispatch_laplace.register(cupy.ndarray)
     def cupy_laplace(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.laplace
+        return cupyx.scipy.ndimage.laplace
 
 
 # ================== prewitt ==================
 @dispatch_prewitt.register(np.ndarray)
 def numpy_prewitt(*args, **kwargs):
-    return scipy.ndimage.filters.prewitt
+    return scipy.ndimage.prewitt
 
 
 @dispatch_prewitt.register_lazy("cupy")
@@ -104,13 +104,13 @@ def register_cupy_prewitt():
 
     @dispatch_prewitt.register(cupy.ndarray)
     def cupy_prewitt(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.prewitt
+        return cupyx.scipy.ndimage.prewitt
 
 
 # ================== sobel ==================
 @dispatch_sobel.register(np.ndarray)
 def numpy_sobel(*args, **kwargs):
-    return scipy.ndimage.filters.sobel
+    return scipy.ndimage.sobel
 
 
 @dispatch_sobel.register_lazy("cupy")
@@ -120,13 +120,13 @@ def register_cupy_sobel():
 
     @dispatch_sobel.register(cupy.ndarray)
     def cupy_sobel(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.sobel
+        return cupyx.scipy.ndimage.sobel
 
 
 # ================== gaussian_filter ==================
 @dispatch_gaussian_filter.register(np.ndarray)
 def numpy_gaussian_filter(*args, **kwargs):
-    return scipy.ndimage.filters.gaussian_filter
+    return scipy.ndimage.gaussian_filter
 
 
 @dispatch_gaussian_filter.register_lazy("cupy")
@@ -136,13 +136,13 @@ def register_cupy_gaussian_filter():
 
     @dispatch_gaussian_filter.register(cupy.ndarray)
     def cupy_gaussian_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.gaussian_filter
+        return cupyx.scipy.ndimage.gaussian_filter
 
 
 # ================== gaussian_gradient_magnitude ==================
 @dispatch_gaussian_gradient_magnitude.register(np.ndarray)
 def numpy_gaussian_gradient_magnitude(*args, **kwargs):
-    return scipy.ndimage.filters.gaussian_gradient_magnitude
+    return scipy.ndimage.gaussian_gradient_magnitude
 
 
 @dispatch_gaussian_gradient_magnitude.register_lazy("cupy")
@@ -152,13 +152,13 @@ def register_cupy_gaussian_gradient_magnitude():
 
     @dispatch_gaussian_gradient_magnitude.register(cupy.ndarray)
     def cupy_gaussian_gradient_magnitude(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.gaussian_gradient_magnitude
+        return cupyx.scipy.ndimage.gaussian_gradient_magnitude
 
 
 # ================== gaussian_laplace ==================
 @dispatch_gaussian_laplace.register(np.ndarray)
 def numpy_gaussian_laplace(*args, **kwargs):
-    return scipy.ndimage.filters.gaussian_laplace
+    return scipy.ndimage.gaussian_laplace
 
 
 @dispatch_gaussian_laplace.register_lazy("cupy")
@@ -168,13 +168,13 @@ def register_cupy_gaussian_laplace():
 
     @dispatch_gaussian_laplace.register(cupy.ndarray)
     def cupy_gaussian_laplace(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.gaussian_laplace
+        return cupyx.scipy.ndimage.gaussian_laplace
 
 
 # ================== generic_filter ==================
 @dispatch_generic_filter.register(np.ndarray)
 def numpy_generic_filter(*args, **kwargs):
-    return scipy.ndimage.filters.generic_filter
+    return scipy.ndimage.generic_filter
 
 
 @dispatch_generic_filter.register_lazy("cupy")
@@ -184,13 +184,13 @@ def register_cupy_generic_filter():
 
     @dispatch_generic_filter.register(cupy.ndarray)
     def cupy_generic_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.generic_filter
+        return cupyx.scipy.ndimage.generic_filter
 
 
 # ================== minimum_filter ==================
 @dispatch_minimum_filter.register(np.ndarray)
 def numpy_minimum_filter(*args, **kwargs):
-    return scipy.ndimage.filters.minimum_filter
+    return scipy.ndimage.minimum_filter
 
 
 @dispatch_minimum_filter.register_lazy("cupy")
@@ -200,13 +200,13 @@ def register_cupy_minimum_filter():
 
     @dispatch_minimum_filter.register(cupy.ndarray)
     def cupy_minimum_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.minimum_filter
+        return cupyx.scipy.ndimage.minimum_filter
 
 
 # ================== median_filter ==================
 @dispatch_median_filter.register(np.ndarray)
 def numpy_median_filter(*args, **kwargs):
-    return scipy.ndimage.filters.median_filter
+    return scipy.ndimage.median_filter
 
 
 @dispatch_median_filter.register_lazy("cupy")
@@ -216,13 +216,13 @@ def register_cupy_median_filter():
 
     @dispatch_median_filter.register(cupy.ndarray)
     def cupy_median_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.median_filter
+        return cupyx.scipy.ndimage.median_filter
 
 
 # ================== maximum_filter ==================
 @dispatch_maximum_filter.register(np.ndarray)
 def numpy_maximum_filter(*args, **kwargs):
-    return scipy.ndimage.filters.maximum_filter
+    return scipy.ndimage.maximum_filter
 
 
 @dispatch_maximum_filter.register_lazy("cupy")
@@ -232,13 +232,13 @@ def register_cupy_maximum_filter():
 
     @dispatch_maximum_filter.register(cupy.ndarray)
     def cupy_maximum_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.maximum_filter
+        return cupyx.scipy.ndimage.maximum_filter
 
 
 # ================== rank_filter ==================
 @dispatch_rank_filter.register(np.ndarray)
 def numpy_rank_filter(*args, **kwargs):
-    return scipy.ndimage.filters.rank_filter
+    return scipy.ndimage.rank_filter
 
 
 @dispatch_rank_filter.register_lazy("cupy")
@@ -248,13 +248,13 @@ def register_cupy_rank_filter():
 
     @dispatch_rank_filter.register(cupy.ndarray)
     def cupy_rank_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.rank_filter
+        return cupyx.scipy.ndimage.rank_filter
 
 
 # ================== percentile_filter ==================
 @dispatch_percentile_filter.register(np.ndarray)
 def numpy_percentile_filter(*args, **kwargs):
-    return scipy.ndimage.filters.percentile_filter
+    return scipy.ndimage.percentile_filter
 
 
 @dispatch_percentile_filter.register_lazy("cupy")
@@ -264,13 +264,13 @@ def register_cupy_percentile_filter():
 
     @dispatch_percentile_filter.register(cupy.ndarray)
     def cupy_percentile_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.percentile_filter
+        return cupyx.scipy.ndimage.percentile_filter
 
 
 # ================== uniform_filter ==================
 @dispatch_uniform_filter.register(np.ndarray)
 def numpy_uniform_filter(*args, **kwargs):
-    return scipy.ndimage.filters.uniform_filter
+    return scipy.ndimage.uniform_filter
 
 
 @dispatch_uniform_filter.register_lazy("cupy")
@@ -280,7 +280,7 @@ def register_cupy_uniform_filter():
 
     @dispatch_uniform_filter.register(cupy.ndarray)
     def cupy_uniform_filter(*args, **kwargs):
-        return cupyx.scipy.ndimage.filters.uniform_filter
+        return cupyx.scipy.ndimage.uniform_filter
 
 
 # ================== threshold_local_mean ==================
