@@ -21,6 +21,7 @@ import dask_image.ndfilters
 @pytest.mark.parametrize(
     "da_func",
     [
+        dask_image.ndfilters.gaussian,
         dask_image.ndfilters.gaussian_filter,
         dask_image.ndfilters.gaussian_gradient_magnitude,
         dask_image.ndfilters.gaussian_laplace,
@@ -49,7 +50,8 @@ def test_gaussian_filters_params(da_func, err_type, sigma, truncate):
 @pytest.mark.parametrize(
     "sp_func, da_func",
     [
-        (scipy.ndimage.filters.gaussian_filter, dask_image.ndfilters.gaussian_filter),  # noqa: E501
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian),
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian_filter),
     ]
 )
 def test_gaussian_filters_identity(sp_func, da_func, order, sigma, truncate):
@@ -76,6 +78,7 @@ def test_gaussian_filters_identity(sp_func, da_func, order, sigma, truncate):
 @pytest.mark.parametrize(
     "da_func",
     [
+        dask_image.ndfilters.gaussian,
         dask_image.ndfilters.gaussian_filter,
         dask_image.ndfilters.gaussian_gradient_magnitude,
         dask_image.ndfilters.gaussian_laplace,
@@ -98,6 +101,7 @@ def test_gaussian_filter_shape_type(da_func):
 @pytest.mark.parametrize(
     "da_func",
     [
+        dask_image.ndfilters.gaussian,
         dask_image.ndfilters.gaussian_filter,
         dask_image.ndfilters.gaussian_gradient_magnitude,
         dask_image.ndfilters.gaussian_laplace,
@@ -131,11 +135,11 @@ def test_gaussian_filter_comprehensions(da_func):
 @pytest.mark.parametrize(
     "sp_func, da_func",
     [
-        (scipy.ndimage.filters.gaussian_filter,
-         dask_image.ndfilters.gaussian_filter),
-        (scipy.ndimage.filters.gaussian_gradient_magnitude,
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian),
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian_filter),
+        (scipy.ndimage.gaussian_gradient_magnitude,
          dask_image.ndfilters.gaussian_gradient_magnitude),
-        (scipy.ndimage.filters.gaussian_laplace,
+        (scipy.ndimage.gaussian_laplace,
          dask_image.ndfilters.gaussian_laplace),
     ]
 )
@@ -176,7 +180,8 @@ def test_gaussian_filters_compare(sp_func, da_func, sigma, truncate):
 @pytest.mark.parametrize(
     "sp_func, da_func",
     [
-        (scipy.ndimage.filters.gaussian_filter, dask_image.ndfilters.gaussian_filter),  # noqa: E501
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian),
+        (scipy.ndimage.gaussian_filter, dask_image.ndfilters.gaussian_filter),
     ]
 )
 def test_gaussian_derivative_filters_compare(sp_func, da_func,
